requirements.txt
django-jsonfield
Django==2.1.5
django-cors-headers==2.4.0
django-filter==2.1.0
django-rest-auth==0.9.3
django-rest-framework==0.1.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
idna==2.6
invoke==1.2.0
py4j==0.10.7
PyMySQL==0.9.3
pyspark==2.3.1
pytz==2018.9
PyYAML==3.12
six==1.11.0
typing==3.5.3.0
unicodecsv==0.14.1
